Fix incorrect byte conversion of integers for hashing
Calling `bytes(value)` is going to create a buffer that is `value` long,
filled with zeros. This was being passed to `hasher(bytes(value))` for
example, the timestamp which meant Python was creating a massive buffer
and getting it hashed into blake2b.

Tests were taking 10s each due to this which has now been fixed with this patch.