Start transitioning away from payment token hash
First accept just the platform specific transaction metadata, get it
working, then see if we can come up with a reasonable abstraction of the
two. If not, it's better not to as we will most likely design the wrong
abstraction for the payment token and that's best to let the platforms
lead as they all work in various different ways that are optimal to
them.