Don't dedupe google purchase token by token
The Google purchase token is reused across billing cycles and perpetuity
until whatever condition. This means that we need to support witnessing
unredeemed payments from Google with the same purchase token.

It's the caller's reponsibility to verify that the token is valid to be
reused by reverifying it with the Google APIs. There's only weak
deduplication using the event timestamp which Google is going to
guarantee as publishing these events in order.