Google exposes an order ID which is fit for uniquely identifying TXs
This can replace the notification ms which was a pre-emptive heuristic
as to what we believed could be used as a unique identifier for
a transaction. Now users have to pass in the purchase token AND the
order ID to disambiguate which particular payment they are redeeming.

The backend only accepts this if the payment was independently witnessed
from Google and subsequently verified using their verification APIs.