Remove activated and set expiry timestamp on payment
The way Google wants to work is they are the authoritative arbiter of
the expiry time that a payment is authorised for. The backend was
designed more to be the arbiter of subscription duration by deciding
when to expire a user's entitlement.

When a user changes their renewal preference on Google this enables/disables the
grace period and so expiry can change for example. It's best to defer to google
as we are required to at the minimum provide entitlement for atleast the
duration that google claims which is best done by matching their durations.