Split up expiry methods into individual methods
There was an issue with apple where refunding the payment marked it as
expired instead of refunded due to expiry methods being used to handle
refunds and expiry. These are now separate concepts.

Apple now for a refund will actually run the refund logic instead of expiry
logic. Thus preserving the refunded state on the transaction