Use TXs for atomic DB changes, simplify error handling and inline functions
- Remove unused backend.verify_google_payment_token_hash function

- Revocation of purchase tokens was failing because we were trying to
revoke by order ID. The order ID grabbed from the notification is
a unique order ID for that event, it has no relationship to the previous
payment we need to revoke. That information that links us back to the
old payment is the linked-payment token itself. This means the API was
fundamentally broken and doing the wrong thing.

This is similar to Apple in that we get an original transaction ID that is
shared across payments. So for Google I've introduced a custom function to
handle the revocation logic.

Fixing this broke some tests that made the wrong assumptions but they are now
fixed. There is still some cleanup to be done now as the old revocation
function is kind of broken and not-fit for purpose.

- Remove the need to constantly parse the pro-plan from the subscription
notification. We hoist up parsing of pro plan from the base plan ID
where we extract subscription data into our local SubscriptionEvent data
structures. By hoisting it up earlier we remove a potentially failure
point in the notification handling code.

This turns some functions in one-liners that call into the backend
layer. These one-liners can then be inlined into the notification
handling function itself removing 1 level from the call stack.

- Inline the combo function of add unredeeemd and add payment function
by using a transaction.

- Remove some impossible match branches

- Make handle_notification be more pure by returning
a UserErrorTransaction instead of taking it by reference