Derive user expiry/grace every time payment is updated
This has some knock-on effect to everything else. We take on the burden
of calculating this for the caller. Previously we gave them a list of payments
and grabbed the expiry time from the latest payment. However recently we changed
to reporting the latest expiry time including grace.

This subtle logic change means that the SQL query to pull out said data
is not as trivial. We implement this logic in python rather than the
query language.

The user record is now always reporting this up-to date value which
means that anything we make some meaningful change to the payment list
of a master public key we need to "recalc" the user's expiry, grace and
auto-renewing status onto their user row. Potentially this could be done
better with a view but that's for a future problem.