Avoid revoking of old payments on new payment
There isn't a need for this as a pro proof is no longer the source of
truth as to the user's current Session Pro plan. This avoids churn in
the revocation list at a user's billing cycle, they can continue to hold
onto their proof until expiry and only then request a new one when
needed.

This is precursor work to automatically redeeming an unredeemed payment
if there was a prior master public key that has claimed the payment
token associated with a payment. Automatic redeeming without this was
going to have that side effect of modifying the reoke list when it
doesn't need to