Functional-style parse and handle of notifications, minimise extra long google names
We separate handling of notification from parsing. This means that
handling can make more assumptions because we parse out the contents
into a more valid representation before attempting to do business logic
on it. It also sets up a framework for having more control over the
pulling-subscription we have with google by being able to parse the
notification upfront and storing it if it needs to be handled later
(e.g. a notification that failed because a retry notification needs to
be executed first before it).

Reduce some of the verbosity of the names in the Google layer they are
overly verbose for little gain.