Override pro details user pro status if revoked
Since we can revoke from the CLI, we cannot just rely on the list of
payments to determine if a user is revoked. We must now also check the
revocation table for the user and return then as expired if that's
the case.

A revocation is not necessarily paired with a payment and actually
allowing the CLI to revoke a payment may cause issues down the line
because it interferes with the payment lifecycle which is largely
govered by notifications from the Google/iOS app store(s) respectively.